requirements.txt
# Core dependencies
numpy>=1.21.0
cryptography>=3.4.8

# Optional dependencies for enhanced functionality
sentence-transformers>=2.2.0  # For semantic embeddings
faiss-cpu>=1.7.0  # For efficient similarity search
opencv-python>=4.5.0  # For video/image processing

# New dependencies for enhanced file format features
aiofiles>=0.8.0      # For async file operations
brotli>=1.0.9        # For Brotli compression
zstandard>=0.18.0    # For Zstandard compression
xxhash>=3.0.0        # For fast hashing
msgpack>=1.0.4       # For efficient serialization
jsonschema>=4.0.0    # For schema validation
click>=8.0.0         # For CLI interface
tqdm>=4.64.0         # For progress bars
psutil>=5.8.0        # For system monitoring
networkx>=2.6.0      # For knowledge graphs
matplotlib>=3.4.0    # For visualization
pillow>=8.3.0        # For image processing

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0.0